add unit test functionality
