+++ PsychicPets.Tests/UnitTest1.cs
@@ -1,0 +1,15 @@
+namespace PsychicPets.Tests;
+
+public class Tests
+{
+    [SetUp]
+    public void Setup()
+    {
+    }
+
+    [Test]
+    public void Test1()
+    {
+        Assert.Pass();
+    }
+}
